Refactor DateRangePicker, and fixed CreateNewDiscountCode for PR
