+++ src/components/common/DateRangePicker.tsx
@@ -6,11 +6,13 @@ import { MomentRange, YYYY_MM_DD } from 'src/utils/dateUtils';
 type DateRangePickerProps = {
   dateRange: MomentRange;
   updateDateRange: (newRange: (oldRange: MomentRange) => MomentRange) => void;
+  canSelectPastDate?: boolean;
 };
 
 const DateRangePicker = ({
   dateRange,
-  updateDateRange
+  updateDateRange,
+  canSelectPastDate = true
 }: DateRangePickerProps) => {
   return (
     <>
@@ -18,7 +20,10 @@ const DateRangePicker = ({
         label='Start Date'
         value={dateRange[0].format(YYYY_MM_DD)}
         minDate={moment('2000-01-01')}
-        shouldDisableDate={(date) => moment(date).isAfter(dateRange[1])}
+        shouldDisableDate={(date) =>
+          moment(date).isAfter(dateRange[1]) ||
+          (!canSelectPastDate && moment(date).isBefore(dateRange[0]))
+        }
         onChange={(date) =>
           updateDateRange((prev) => [moment(date).startOf('day'), prev[1]])
         }
